Add files via upload
We recreated the game of life on a game board. Using nested loops we created three functions, where we counted the number of live neighbors, updated the game board to the next step using counters that incremented every time we went through the nested loop. Lastly we checked the number of cells alive by looking at the spaces near the current one.